+++ api/v1alpha1/vaultsecret_types.go
@@ -68,6 +68,8 @@ type VaultSecretSpec struct {
 // VaultSecretStatus defines the observed state of VaultSecret
 type VaultSecretStatus struct {
 	Conditions []metav1.Condition `json:"conditions,omitempty"`
+	Expires    bool               `json:"expires"`
+	ExpiresAt  string             `json:"expiresAt,omitempty"`
 }
 
 // +kubebuilder:object:root=true
