+++ vault/database.go
@@ -2,26 +2,29 @@ package vault
 
 import (
 	"fmt"
-	"github.com/go-logr/logr"
+	"github.com/hashicorp/vault/api"
 	"time"
 )
 
-func (c *Client) GetDatabaseCreds(path string, role string, mylog logr.Logger) (map[string][]byte, *time.Time, error) {
-	r, err := c.client.Logical().ReadWithData(path+"/creds/"+role, map[string][]string{})
+func (c *Client) GetDatabaseCreds(path string, role string) (*api.Secret, *time.Time, error) {
+	secret, err := c.client.Logical().ReadWithData(path+"/creds/"+role, map[string][]string{})
 	if err != nil {
 		return nil, nil, err
 	}
 
-	if r == nil {
+	if secret == nil {
 		return nil, nil, fmt.Errorf("database credentials is nil")
 	}
 
-	data, err := convertData(r.Data, []string{"username", "password"}, false)
 	if err != nil {
 		return nil, nil, err
 	}
 
-	expiration := time.Now().Add(time.Duration(r.LeaseDuration) * time.Second)
+	expiresAt := time.Now().Add(time.Duration(secret.LeaseDuration) * time.Second)
 
-	return data, &expiration, nil
+	return secret, &expiresAt, nil
+}
+
+func (c *Client) DatabaseRenderData(secret *api.Secret) (map[string][]byte, error) {
+	return convertData(secret.Data, []string{"username", "password"}, false)
 }
